requirements.txt
discord.py
uvloop
black
python-dotenv
motor[srv]
cprint
colorama